pyproject.toml
[project]
name = "gym-api"
version = "0.1.0"
description = "Production-ready Gym SaaS API with FastAPI, PostgreSQL and Docker."
requires-python = ">=3.11"
dependencies = [
  "alembic>=1.14.0",
  "asyncpg>=0.30.0",
  "celery>=5.4.0",
  "email-validator>=2.2.0",
  "fastapi>=0.116.0",
  "greenlet>=3.1.1",
  "httpx>=0.28.1",
  "passlib[argon2]>=1.7.4",
  "psycopg[binary]>=3.2.5",
  "pydantic-settings>=2.7.1",
  "python-jose[cryptography]>=3.3.0",
  "python-multipart>=0.0.20",
  "redis>=5.2.1",
  "sqlalchemy>=2.0.36",
  "structlog>=25.1.0",
  "uvicorn[standard]>=0.34.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.4",
  "pytest-asyncio>=0.25.0",
  "ruff>=0.9.3",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
asyncio_mode = "auto"

[build-system]
requires = ["setuptools>=69.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
include = ["app*"]
